+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## v0.9.0-rc18 [Unreleased]
+
+### Bugfixes
+- [#2100](https://github.com/influxdb/influxdb/pull/2100): Use channel to synchronize collectd shutdown.
+- [#2100](https://github.com/influxdb/influxdb/pull/2100): Synchronize access to shard index.
+
 ## v0.9.0-rc17 [2015-03-29]
 
 ### Features